+++ tests/integration/kraken_exchange/README.md
@@ -9,9 +9,3 @@ Notable details:
   orders and trades locally.
 - The tests are parameterized to run against multiple symbols (e.g. BTCUSD,
   AAPLxUSD) for each strategy.
-
-## Future Work
-
-- Outsource common code to reduce duplication.
-- The KrakenTestManager could be derived from a base test manager class to
-  provide common functionality for all exchange tests.
