+++ tests/integration/kraken_exchange/test_kraken_cdca_refactored.py
@@ -1,0 +1,30 @@
+# -*- mode: python; coding: utf-8 -*-
+#
+# Copyright (C) 2025 Benjamin Thomas Schwertfeger
+# All rights reserved.
+# https://github.com/btschwertfeger
+#
+
+"""Integration tests for cDCA strategy on Kraken exchange."""
+
+import logging
+from typing import Callable
+from unittest import mock
+
+import pytest
+
+LOG = logging.getLogger(__name__)
+
+@pytest.mark.integration
+@pytest.mark.asyncio
+@mock.patch("infinity_grid.adapters.exchanges.kraken.sleep", return_value=None)
+@mock.patch("infinity_grid.strategies.grid_base.sleep", return_value=None)
+async def test_kraken_cdca(
+    mock_sleep1: mock.MagicMock,  # noqa: ARG001
+    mock_sleep2: mock.MagicMock,  # noqa: ARG001
+    caplog: pytest.LogCaptureFixture,
+    test_manager_factory: Callable,
+) -> None:
+    caplog.set_level(logging.INFO)
+
+    pass
