+++ tests/integration/kraken_exchange/test_kraken_swing.py
@@ -14,7 +14,7 @@ from unittest import mock
 
 import pytest
 
-from .helper import KrakenTestManager
+from .kraken_test_manager import KrakenIntegrationTestManager
 
 LOG = logging.getLogger(__name__)
 
@@ -270,7 +270,7 @@ async def test_kraken_swing(
     mock_sleep2: mock.MagicMock,  # noqa: ARG001
     mock_sleep3: mock.MagicMock,  # noqa: ARG001
     caplog: pytest.LogCaptureFixture,
-    kraken_test_manager_factory: Callable[[str, str], KrakenTestManager],
+    test_manager_factory: Callable[[str, str], KrakenIntegrationTestManager],
     symbol: str,
     expectations: dict,
 ) -> None:
@@ -281,7 +281,7 @@ async def test_kraken_swing(
     LOG.info("******* Starting SWING integration test *******")
     caplog.set_level(logging.INFO)
 
-    tm = kraken_test_manager_factory(symbol, strategy="SWING")
+    tm = test_manager_factory("Kraken", symbol, strategy="SWING")
     await tm.initialize_engine()
     await tm.trigger_prepare_for_trading(initial_ticker=expectations["initial_ticker"])
 
@@ -485,7 +485,7 @@ async def test_kraken_swing_unfilled_surplus(
     mock_sleep2: mock.Mock,  # noqa: ARG001
     mock_sleep3: mock.Mock,  # noqa: ARG001
     caplog: pytest.LogCaptureFixture,
-    kraken_test_manager_factory: Callable[[str, str], KrakenTestManager],
+    test_manager_factory: Callable[[str, str], KrakenIntegrationTestManager],
     symbol: str,
     expectations: dict,
 ) -> None:
@@ -501,7 +501,7 @@ async def test_kraken_swing_unfilled_surplus(
     LOG.info("******* Starting SWING unfilled surplus integration test *******")
     caplog.set_level(logging.INFO)
 
-    tm = kraken_test_manager_factory(symbol, strategy="SWING")
+    tm = test_manager_factory("Kraken", symbol, strategy="SWING")
     await tm.initialize_engine()
     await tm.trigger_prepare_for_trading(initial_ticker=expectations["initial_ticker"])
